flask-backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
flask-jwt-extended==4.7.1
pymongo==4.16.0
python-dotenv==1.0.0
bcrypt==5.0.0
stripe==15.0.1
marshmallow==4.3.0
pytest==9.0.3
pytest-flask==1.3.0
mongomock==4.3.0
python-dateutil==2.9.0.post0
gunicorn==25.3.0